// repositorio/SIMTR/4-implementacao/simtr-web/src/app/cruds/tipo-relacionamento/pesquisa/model/grid-tipo-relacionamento.ts
export class GridTipoRelacionamento {
    id: number;
    nome: string;
    tipo_pessoa: string;
    indicador_principal: boolean;
    indicador_relacionado: boolean;
    indicador_sequencia: boolean;
    indicador_receita_pf: boolean;
    indicador_receita_pj: boolean;
    ultima_alteracao: string;
}